+++ data/db/db_engine.py
@@ -7,6 +7,7 @@ from sqlalchemy.orm import sessionmaker
 import inspect
 
 from dotenv import load_dotenv
+
 load_dotenv()
 
 POSTGRES_USER = os.environ.get("POSTGRES_USER", "user")
@@ -15,7 +16,7 @@ POSTGRES_DB = os.environ.get("POSTGRES_DB", "bcs_database")
 POSTGRES_HOST = os.environ.get("POSTGRES_HOST", "localhost")
 POSTGRES_PORT = os.environ.get("POSTGRES_PORT", "5432")
 
-DB_URL = f'postgresql+psycopg2://{POSTGRES_USER}:{POSTGRES_PASSWORD}@{POSTGRES_HOST}:{POSTGRES_PORT}/{POSTGRES_DB}'
+DB_URL = f"postgresql+psycopg2://{POSTGRES_USER}:{POSTGRES_PASSWORD}@{POSTGRES_HOST}:{POSTGRES_PORT}/{POSTGRES_DB}"
 
 engine = create_engine(DB_URL)
 
@@ -27,8 +28,8 @@ Session = sessionmaker(bind=engine)
 
 # Define a model for your data
 class User(Base):
-    __tablename__ = 'users'
-    __table_args__ = {'schema': 'BCS'}
+    __tablename__ = "users"
+    __table_args__ = {"schema": "BCS"}
 
     id = Column(Integer, primary_key=True)
     name = Column(String)
@@ -36,7 +37,17 @@ class User(Base):
     password = Column(String)
 
 
-from data.model import action, blob, calendar, event, event_type, league, result, sim_data, name_suggestion  # noqa: F401, E402
+from data.model import (
+    action,
+    blob,
+    calendar,
+    event,
+    event_type,
+    league,
+    result,
+    sim_data,
+    name_suggestion,
+)  # noqa: F401, E402
 
 
 def create_session():
@@ -57,22 +68,27 @@ def transaction_scope():
 
 
 def transactional(func):
-    """ Decorator that creates a new session if one is not provided """
+    """Decorator that creates a new session if one is not provided"""
 
     def is_session_in_args(args):
         sig = inspect.signature(func)
         params = list(sig.parameters.keys())
-        if 'session' in params:
-            session_index = params.index('session')
+        if "session" in params:
+            session_index = params.index("session")
             return len(args) > session_index and args[session_index] is not None
 
     @wraps(func)
     def wrapper(*args, **kwargs):
-        if 'session' in kwargs and kwargs['session'] is not None or is_session_in_args(args):
+        if (
+            "session" in kwargs
+            and kwargs["session"] is not None
+            or is_session_in_args(args)
+        ):
             # Use the existing session
             return func(*args, **kwargs)
         else:
             # Create a new session
             with transaction_scope() as session:
                 return func(*args, session=session, **kwargs)
+
     return wrapper